+++ Changelog.md
@@ -3,6 +3,7 @@
 Features:
  * Parser: Display previous visibility specifier in error if multiple are found.
  * Syntax Checker: Support ``pragma experimental <feature>;`` to turn on experimental features.
+ * Static Analyzer: Warn about large storage structures.
  * Metadata: Store experimental flag in metadata CBOR.
 
 Bugfixes: